Add Streamable objects
Item, Sprint, Ticket and User are now Streamable.
This means the treams of lists like from getUsersStream will only update
if a new object is added or removed otherwise every object has its own
stream that will only update if that object was changed. This also means
that no longer a new version of each object is created for every stream
but instead there is only one object that gets updated when there is
changed data from the API.
Adjsuted all places where it is used so it still works but the
per object streams are not yet used in many situations.

The User no longer has an avatar property but instead an async
function to get it.
Added functions to get a specific and the current user/user stream.
Added functions to reload data without using it to CachedAPI.
Also some refactoring related to the streams.